requirements.txt
# Discord
discord.py>=2.3.2

# Database
aiosqlite>=0.19.0
SQLAlchemy>=2.0.23
alembic>=1.13.0
greenlet>=3.0.0

# Cache & Rate Limiting
redis>=5.0.1
aioredis>=2.0.1

# Monitoring
prometheus-client>=0.19.0

# Logging
structlog>=23.2.0
python-json-logger>=2.0.7

# Utilities
python-dotenv>=1.0.0
asyncio>=3.4.3

# Development
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
isort>=5.12.0
mypy>=1.7.0
